#75, PNG encoder fails to handle 8bit gray+alpha images

+++ library/geocore/src/test/java/it/geosolutions/imageio/pam/PAMParserTest.java
@@ -31,7 +31,7 @@ public class PAMParserTest {
 
         // Scan the first band
         final PAMRasterBand band = bands.get(0);
-        assertEquals(1, band.getBand());
+        assertEquals(1, (int) band.getBand());
 
         // Checking histogram properties
         final Histograms histograms = band.getHistograms();
